Move lcms related code from lib to src/shared
